fix: slideshow not playing in fullscreen mode with zoom plugin
