+++ src/plugins/slideshow/Slideshow.tsx
@@ -1,8 +1,9 @@
 import * as React from "react";
 
 import { Plugin } from "../../types.js";
+import { createModule, MODULE_CONTROLLER, PLUGIN_SLIDESHOW } from "../../core/index.js";
+import { SlideshowContextProvider } from "./SlideshowContext.js";
 import { SlideshowButton } from "./SlideshowButton.js";
-import { PLUGIN_SLIDESHOW } from "../../core/index.js";
 
 export const defaultSlideshowProps = {
     autoplay: false,
@@ -10,16 +11,12 @@ export const defaultSlideshowProps = {
 };
 
 /** Slideshow plugin */
-export const Slideshow: Plugin = ({ augment }) => {
-    augment(({ slideshow: originalSlideshow, toolbar: { buttons, ...restToolbar }, ...restProps }) => ({
-        toolbar: {
-            buttons: [<SlideshowButton key={PLUGIN_SLIDESHOW} />, ...buttons],
-            ...restToolbar,
-        },
-        slideshow: {
-            ...defaultSlideshowProps,
-            ...originalSlideshow,
-        },
+export const Slideshow: Plugin = ({ augment, append }) => {
+    augment(({ slideshow: slideshowProps, toolbar: { buttons, ...restToolbar }, ...restProps }) => ({
+        toolbar: { buttons: [<SlideshowButton key={PLUGIN_SLIDESHOW} />, ...buttons], ...restToolbar },
+        slideshow: { ...defaultSlideshowProps, ...slideshowProps },
         ...restProps,
     }));
+
+    append(MODULE_CONTROLLER, createModule(PLUGIN_SLIDESHOW, SlideshowContextProvider));
 };
