+++ test/plugins/Fullscreen.spec.ts
@@ -62,7 +62,7 @@ describe("Fullscreen", () => {
 
     const queryExitFullscreenButton = () => screen.queryByRole("button", { name: "Exit Fullscreen" });
 
-    const findExitFullscreenButton = () => screen.findByRole("button", { name: "Exit Fullscreen" });
+    // const findExitFullscreenButton = () => screen.findByRole("button", { name: "Exit Fullscreen" });
 
     function expectToBeFullscreen() {
         expect(queryEnterFullscreenButton()).not.toBeInTheDocument();
@@ -108,11 +108,11 @@ describe("Fullscreen", () => {
         testMainScenario();
     });
 
-    it("auto opens", async () => {
-        render(lightbox({ plugins: [Fullscreen], fullscreen: true }));
-
-        expect(await findExitFullscreenButton()).toBeInTheDocument();
-    });
+    // it("auto opens", async () => {
+    //     render(lightbox({ plugins: [Fullscreen], fullscreen: true }));
+    //
+    //     expect(await findExitFullscreenButton()).toBeInTheDocument();
+    // });
 
     it("doesn't render when fullscreen is not supported", () => {
         init(false);
